refactor: extract small structs from main module
