+++ src/gestures/meson.build
@@ -2,6 +2,8 @@ gestures = static_library('gestures',
   'Gestures.cpp',
   'Shared.cpp',
   'Actions.cpp',
+  'CompletedGesture.cpp',
+  'DragGesture.cpp',
   dependencies: [
     wftouch,
   ])
